Kernel update from 3.4.7 to 8

+++ fs/locks.c
@@ -184,7 +184,7 @@ static int flock_make_lock(struct file *filp, struct file_lock **lock,
 	return 0;
 }
 
-static int assign_type(struct file_lock *fl, int type)
+static int assign_type(struct file_lock *fl, long type)
 {
 	switch (type) {
 	case F_RDLCK:
@@ -312,7 +312,7 @@ static const struct lock_manager_operations lease_manager_ops = {
 	.lm_change = lease_modify,
 };
 
-static int lease_init(struct file *filp, int type, struct file_lock *fl)
+static int lease_init(struct file *filp, long type, struct file_lock *fl)
  {
 	if (assign_type(fl, type) != 0)
 		return -EINVAL;
@@ -329,7 +329,7 @@ static int lease_init(struct file *filp, int type, struct file_lock *fl)
 	return 0;
 }
 
-static struct file_lock *lease_alloc(struct file *filp, int type)
+static struct file_lock *lease_alloc(struct file *filp, long type)
 {
 	struct file_lock *fl = locks_alloc_lock();
 	int error = -ENOMEM;
